+++ packages/htmlembed/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @embeddedchat/htmlembed
 
+## 0.0.5
+
+### Patch Changes
+
+- 14117c0: Fixed: invalid storing of authenticatedUserId
+- Updated dependencies [14117c0]
+  - @embeddedchat/react@0.1.12
+
+## 0.0.4
+
+### Patch Changes
+
+- Updated dependencies [ce5a26b]
+  - @embeddedchat/react@0.1.11
+
 ## 0.0.3
 
 ### Patch Changes